+++ src/dipdup/indexes/evm_subsquid_events/fetcher.py
@@ -6,6 +6,7 @@ from dipdup.datasources.evm_node import EvmNodeDatasource
 from dipdup.datasources.evm_subsquid import SubsquidDatasource
 from dipdup.fetcher import DataFetcher
 from dipdup.fetcher import readahead_by_level
+from dipdup.indexes.evm_node import EVM_NODE_READAHEAD_LIMIT
 from dipdup.indexes.evm_node import MIN_BATCH_SIZE
 from dipdup.indexes.evm_node import EvmNodeFetcher
 from dipdup.indexes.evm_subsquid import SUBSQUID_READAHEAD_LIMIT
@@ -41,7 +42,7 @@ class EvmNodeEventFetcher(EvmNodeFetcher[EvmNodeLogData]):
 
     async def fetch_by_level(self) -> AsyncIterator[tuple[int, tuple[EvmNodeLogData, ...]]]:
         event_iter = self._fetch_by_level()
-        async for level, batch in readahead_by_level(event_iter, limit=SUBSQUID_READAHEAD_LIMIT):
+        async for level, batch in readahead_by_level(event_iter, limit=EVM_NODE_READAHEAD_LIMIT):
             yield level, batch
 
     async def _fetch_by_level(self) -> AsyncIterator[tuple[EvmNodeLogData, ...]]: