+++ src/dipdup/indexes/evm_subsquid_transactions/fetcher.py
@@ -1,9 +1,13 @@
+import random
+import time
 from collections.abc import AsyncIterator
 
 from dipdup.datasources.evm_node import EvmNodeDatasource
 from dipdup.datasources.evm_subsquid import SubsquidDatasource
 from dipdup.fetcher import DataFetcher
 from dipdup.fetcher import readahead_by_level
+from dipdup.indexes.evm_node import EVM_NODE_READAHEAD_LIMIT
+from dipdup.indexes.evm_node import MIN_BATCH_SIZE
 from dipdup.indexes.evm_node import EvmNodeFetcher
 from dipdup.indexes.evm_subsquid import SUBSQUID_READAHEAD_LIMIT
 from dipdup.models.evm_node import EvmNodeTransactionData
@@ -39,12 +43,25 @@ class SubsquidTransactionFetcher(DataFetcher[SubsquidTransactionData]):
 class EvmNodeTransactionFetcher(EvmNodeFetcher[EvmNodeTransactionData]):
     _datasource: EvmNodeDatasource
 
+    async def fetch_by_level(self) -> AsyncIterator[tuple[int, tuple[EvmNodeTransactionData, ...]]]:
+        transaction_iter = self._fetch_by_level()
+        async for level, batch in readahead_by_level(transaction_iter, limit=EVM_NODE_READAHEAD_LIMIT):
+            yield level, batch
+
     async def _fetch_by_level(self) -> AsyncIterator[tuple[EvmNodeTransactionData, ...]]:
+        batch_size = MIN_BATCH_SIZE
         batch_first_level = self._first_level
+        ratelimited: bool = False
+
         while batch_first_level <= self._last_level:
-            node = self.get_random_node()
+            node = random.choice(self._datasources)
+            batch_size = self.get_next_batch_size(batch_size, ratelimited)
+            ratelimited = False
+
+            started = time.time()
+
             batch_last_level = min(
-                batch_first_level + node._http_config.batch_size,
+                batch_first_level + batch_size,
                 self._last_level,
             )
 
@@ -53,18 +70,21 @@ class EvmNodeTransactionFetcher(EvmNodeFetcher[EvmNodeTransactionData]):
                 full_transactions=True,
                 node=node,
             )
+            blocks = sorted(block_batch.values(), key=lambda block: int(block['number'], 16))
 
-            for _, block in sorted(block_batch.items()):
+            finished = time.time()
+            if finished - started >= node._http_config.ratelimit_sleep:
+                ratelimited = True
+
+            for block in blocks:
                 timestamp = int(block['timestamp'], 16)
-                parsed_transactions = tuple(
+                if not block['transactions']:
+                    continue
+
+                parsed_level_transactions = tuple(
                     EvmNodeTransactionData.from_json(transaction, timestamp) for transaction in block['transactions']
                 )
 
-                yield parsed_transactions
+                yield parsed_level_transactions
 
             batch_first_level = batch_last_level + 1
-
-    async def fetch_by_level(self) -> AsyncIterator[tuple[int, tuple[EvmNodeTransactionData, ...]]]:
-        transaction_iter = self._fetch_by_level()
-        async for level, batch in readahead_by_level(transaction_iter, limit=SUBSQUID_READAHEAD_LIMIT):
-            yield level, batch