+++ src/dipdup/indexes/tezos_tzkt_events/fetcher.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING
 
 from dipdup.fetcher import DataFetcher
 from dipdup.fetcher import readahead_by_level
+from dipdup.indexes.tezos_tzkt import TZKT_READAHEAD_LIMIT
 from dipdup.models.tezos_tzkt import TzktEventData
 
 if TYPE_CHECKING:
@@ -14,8 +15,6 @@ if TYPE_CHECKING:
 
 
 class EventFetcher(DataFetcher[TzktEventData]):
-    """Fetches contract events from REST API, merges them and yields by level."""
-
     _datasource: TzktDatasource
 
     def __init__(
@@ -34,15 +33,11 @@ class EventFetcher(DataFetcher[TzktEventData]):
         self._event_tags = event_tags
 
     async def fetch_by_level(self) -> AsyncGenerator[tuple[int, tuple[TzktEventData, ...]], None]:
-        """Iterate over events fetched fetched from REST.
-
-        Resulting data is splitted by level, deduped, sorted and ready to be processed by TzktEventsIndex.
-        """
         event_iter = self._datasource.iter_events(
             self._event_addresses,
             self._event_tags,
             self._first_level,
             self._last_level,
         )
-        async for level, batch in readahead_by_level(event_iter, limit=5_000):
+        async for level, batch in readahead_by_level(event_iter, limit=TZKT_READAHEAD_LIMIT):
             yield level, batch