+++ src/dipdup/indexes/tezos_tzkt_operations/matcher.py
@@ -51,7 +51,8 @@ def prepare_operation_handler_args(
 ) -> deque[OperationsHandlerArgumentU]:
     """Prepare handler arguments, parse parameter and storage."""
     args: deque[OperationsHandlerArgumentU] = deque()
-    for pattern_config, operation_data in zip(handler_config.pattern, matched_operations, strict=True):
+    # NOTE: There can be more pattern items than matched operations; some of them are optional.
+    for pattern_config, operation_data in zip(handler_config.pattern, matched_operations, strict=False):
         if operation_data is None:
             args.append(None)