+++ src/dipdup/indexes/tezos_tzkt.py
@@ -8,7 +8,7 @@ from dipdup.index import IndexConfigT
 from dipdup.index import IndexQueueItemT
 from dipdup.models.tezos_tzkt import TzktMessageType
 
-TZKT_READAHEAD_LIMIT = 5000
+TZKT_READAHEAD_LIMIT = 10000
 
 
 class TzktIndex(