+++ CHANGELOG.md
@@ -9,10 +9,11 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 ### Fixed
 
 - evm.node: Fixed default ratelimit sleep time being too high.
+- evm.subsquid.transactions: Fixed issue with `node_only` flag ignored.
 
 ### Performance
 
-- evm.subsquid.events: Dynamically adjust the batch size when syncing with node.
+- evm.subsquid: Dynamically adjust the batch size when syncing with node.
 
 ## [7.5.0] - 2024-03-08