+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
-## [Unreleased]
+## [7.5.2] - 2024-03-20
 
 ### Fixed
 
@@ -1391,7 +1391,8 @@ This release contains no changes except for the version number.
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/dipdup-io/dipdup/compare/7.5.1...HEAD
+[Unreleased]: https://github.com/dipdup-io/dipdup/compare/7.5.2...HEAD
+[7.5.2]: https://github.com/dipdup-io/dipdup/compare/7.5.1...7.5.2
 [7.5.1]: https://github.com/dipdup-io/dipdup/compare/7.5.0...7.5.1
 [7.5.0]: https://github.com/dipdup-io/dipdup/compare/7.4.0...7.5.0
 [6.5.16]: https://github.com/dipdup-io/dipdup/compare/6.5.15...6.5.16